refactor(harness-bench): split shared server into its own module; hoist imports
Readability/structure cleanup requested in review, no behavior change.

- Split full_server.py out of full_server_driver.py: the server+runner
  lifecycle and agent/session registration (SharedFullServer + spawn/wait/
  config helpers + the shared _find_free_port/_mint_bearer/spawn_omnigent_server
  that native-tui also uses) now live in full_server.py; full_server_driver.py
  keeps just FullServerDriver and its probe/item-scan helpers. Clear seam:
  "the server" vs "the driver that runs probes against it".
- Hoist function-body imports to module top across the package (Any, shutil,
  cli_unavailable_reason, omnigent.harness_capabilities/plugins, LineSink,
  SharedFullServer, socket/io/tarfile/yaml). The only inline imports left are
  intentional and now commented: the optional `rich` dependency (richreport +
  its lazy load in __main__) and two documented cycle-avoidance imports
  (transport→drivers, profile→manifest).
- Update consumers (native_tui_driver, bench) to import the shared helpers
  from full_server; fix the shared-server test to patch bench's namespace
  (bench now imports SharedFullServer at top).

Offline suite 56 passed / 14 skipped, ruff clean, no import cycle.
